requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
python-dotenv>=1.0
Flask-WTF>=1.2
email-validator>=2.1
pytest>=8.0
waitress>=3.0
black>=24.0
ruff>=0.6
Flask-Login>=0.6
